# feat: init start with one cmd