--- a/package.json
+++ b/package.json
@@ -75,9 +75,11 @@
     "lodash.clone": "^4.5.0",
     "lodash.defaults": "^4.2.0",
     "lodash.defaultsdeep": "^4.6.1",
+    "merge-options": "^1.0.1",
     "multiaddr": "^6.1.0",
     "safe-json-stringify": "^1.2.0",
-    "superagent": "^5.0.5"
+    "superagent": "^5.0.5",
+    "temp-write": "^4.0.0"
   },
   "devDependencies": {
     "aegir": "^20.0.0",
@@ -87,7 +89,7 @@
     "dirty-chai": "^2.0.1",
     "go-ipfs-dep": "~0.4.22",
     "husky": "^3.0.4",
-    "ipfs": "~0.37.1",
+    "ipfs": "github:ipfs/js-ipfs#feat/daemon-start-init",
     "is-running": "^2.1.0",
     "lint-staged": "^9.2.5",
     "proxyquire": "^2.1.3",