--- a/src/utils/config-file.js
+++ b/src/utils/config-file.js
@@ -1,0 +1,11 @@
+'use strict'
+const tempWrite = require('temp-write')
+const merge = require('merge-options')
+
+function configFile (type, config) {
+  return tempWrite(JSON.stringify(merge({}, config)), 'config.json')
+}
+
+module.exports = {
+  configFile
+}