ctl timings

before: 

    512 x 0.29 ops/sec ±7.81% (6 runs sampled)
    1024 x 0.30 ops/sec ±3.50% (6 runs sampled)
    2048 x 0.29 ops/sec ±3.97% (6 runs sampled)

after:

    512 x 0.68 ops/sec ±7.11% (8 runs sampled)
    1024 x 0.71 ops/sec ±2.70% (8 runs sampled)
    2048 x 0.69 ops/sec ±3.24% (8 runs sampled)

needs:
https://github.com/ipfs/js-ipfs/pull/2428